+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+### Changed
+- Removed `-reuse` deprecation warning.
+
 ## [0.7.0] - 2016-07-21
 ### Added
 - We now output min and max latency numbers bracketing the latency percentiles. (Fixes #13)